+++ libraries/exoplayer_ima/src/test/java/androidx/media3/exoplayer/ima/FakeExoPlayer.java
@@ -211,11 +211,6 @@ import androidx.media3.test.utils.StubExoPlayer;
   // ExoPlayer methods. Other methods are unsupported.
 
   @Override
-  public AudioComponent getAudioComponent() {
-    return null;
-  }
-
-  @Override
   public Looper getApplicationLooper() {
     return Looper.getMainLooper();
   }